refactored visualization code structure
